fix sim_cosmo.py

+++ notebooks/ex_cosmology/sim_cosmology.py
@@ -21,9 +21,9 @@ from dset import get_dataloader
 parser = argparse.ArgumentParser(description='Cosmology Example')
 parser.add_argument('--batch_size', type=int, default=64, metavar='N',
                     help='input batch size for training (default: 64)')
-parser.add_argument('--init_level', type=int, default=1, metavar='N',
+parser.add_argument('--init_level', type=float, default=1, metavar='N',
                     help='initialization parameter')
-parser.add_argument('--noise_level', type=int, default=0.1, metavar='N',
+parser.add_argument('--noise_level', type=float, default=0.1, metavar='N',
                     help='initialization parameter')
 parser.add_argument('--num_epochs', type=int, default=50, metavar='N',
                     help='number of epochs to train (default: 50)')
